:sparkles: Add tournament progression and team evaluation
This update includes the creation of the `MultiProgression` model and modifications in the `Rozlos` template to handle tournament progression. It also introduces new methods in the `Team` and `Game` models to facilitate the evaluation and scoring of teams within the Tournament module. Lastly, it also includes enhancements to better handle the syncing of upcoming games and tournament settings.